requirements.txt
alabaster==0.7.16
annotated-types==0.7.0
autodoc_pydantic==2.2.0
azure-core==1.31.0
azure-identity==1.18.0
azure-keyvault-secrets==4.8.0
azure-storage-blob==12.23.1
babel==2.16.0
bcrypt==4.2.0
black==24.8.0
certifi==2024.8.30
cffi==1.17.1
charset-normalizer==3.3.2
click==8.1.7
colorama==0.4.6
coverage==7.6.1
cryptography==43.0.1
docutils==0.20.1
idna==3.10
imagesize==1.4.1
iniconfig==2.0.0
isodate==0.6.1
isort==5.13.2
Jinja2==3.1.4
markdown-it-py==3.0.0
MarkupSafe==2.1.5
mdit-py-plugins==0.4.2
mdurl==0.1.2
msal==1.31.0
msal-extensions==1.2.0
mypy-extensions==1.0.0
packaging==24.1
paramiko==3.5.0
pathspec==0.12.1
platformdirs==4.3.6
pluggy==1.5.0
portalocker==2.10.1
pycparser==2.22
pydantic==2.9.2
pydantic-settings==2.5.2
pydantic_core==2.23.4
Pygments==2.18.0
PyJWT==2.9.0
PyNaCl==1.5.0
pytest==8.3.3
pytest-cov==5.0.0
python-dotenv==1.0.1
pywin32==306; platform_system == "Windows"
PyYAML==6.0.2
requests==2.32.3
sentry-sdk==2.15.0
six==1.16.0
snowballstemmer==2.2.0
Sphinx==7.4.7
sphinx-rtd-theme==2.0.0
sphinxcontrib-applehelp==2.0.0
sphinxcontrib-devhelp==2.0.0
sphinxcontrib-htmlhelp==2.1.0
sphinxcontrib-jquery==4.1
sphinxcontrib-jsmath==1.0.1
sphinxcontrib-qthelp==2.0.0
sphinxcontrib-serializinghtml==2.0.0
tqdm==4.66.5
typing_extensions==4.12.2
urllib3==2.2.3
